pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hydroburn"
version = "1.0.0"
description = "Post-wildfire runoff analysis framework"
authors = [
  { name="HydroBurn Agent", email="agent@example.com" },
]
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "geopandas>=0.13.0",
    "shapely>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "markdown>=3.4.0",
]
requires-python = ">=3.10"

[project.scripts]
hydroburn = "hydroburn.cli:cli"
